refactor: files to new folder structure and rename ULSReader to MacOSReader

+++ src/modules/logcollector/src/macos_reader/oslogstore/tests/oslogstore_tests.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include <oslogstore_wrapper.hpp>
+#include <oslogstore.hpp>
 
 #include <chrono>
 
@@ -13,7 +13,7 @@ TEST(OSLogStoreWrapperTest, Constructor)
 TEST(OSLogStoreWrapperTest, AllEntriesDoesNotThrow)
 {
     OSLogStoreWrapper oslogstore;
-    EXPECT_NO_THROW(oslogstore.AllEntries(std::chrono::system_clock::now().time_since_epoch().count(),
+    EXPECT_NO_THROW(oslogstore.AllEntries(static_cast<double>(std::chrono::system_clock::now().time_since_epoch().count()),
                                           "process != \"wazuh-agent\"",
                                           OSLogStoreWrapper::LogLevel::Info));
 }
@@ -28,7 +28,7 @@ TEST(OSLogStoreWrapperTest, AllEntiresThrowForInvalidStartTimes)
 TEST(OSLogStoreWrapperTest, AllEntriesThrowsWhenGivenAnInvalidQuery)
 {
     OSLogStoreWrapper oslogstore;
-    EXPECT_ANY_THROW(oslogstore.AllEntries(std::chrono::system_clock::now().time_since_epoch().count(),
+    EXPECT_ANY_THROW(oslogstore.AllEntries(static_cast<double>(std::chrono::system_clock::now().time_since_epoch().count()),
                                            "invalidfieldthatwillthrow != \"wazuh-agent\"",
                                            OSLogStoreWrapper::LogLevel::Info));
 }
